Anomaly Detection & Root Cause Analysis
In Data Centers, once an anomaly is detected, the main challenge of Site Reliability Engineers is to identify and understand the root cause of that anomaly as soon as possible. The task of root cause analysis (RCA) is generally done manually, and as a result, takes significant time and effort. The objective of this project was to reduce the human effort by automating anomaly detection (AD) and RCA, using state-of-the-art Machine Learning and Causal Inference methods.